# packages/shared/src/types/prisma.types.ts
// This file was auto-generated by prisma-generator-typescript-interfaces

export type UnitType = 'APARTMENT' | 'OFFICE' | 'COMMERCIAL'

export type UnitStatus = 'AVAILABLE' | 'OCCUPIED' | 'MAINTENANCE'

export type StaffRole = 'MANAGER' | 'SECURITY' | 'MAINTENANCE' | 'OTHER'

export type PropertyType = 'RESIDENTIAL' | 'COMMERCIAL' | 'MIXED'

export type UserRole = 'ADMIN' | 'MANAGER' | 'RESIDENT' | 'STAFF'

export type Building = {
  id: number
  name: string
  description: string | null
  yearBuilt: number
  propertyType: PropertyType
  address: string
  district: string
  city: string
  province: string
  postalCode: string | null
  floors: number
  phoneNumber: string | null
  email: string | null
  isActive: boolean
  createdAt: Date
  updatedAt: Date
}

export type User = {
  id: number
  email: string
  password: string
  role: UserRole
  firstName: string
  lastName: string
  createdAt: Date
  updatedAt: Date
  Staff?: Staff[]
}

export type Unit = {
  id: number
  buildingId: number
  number: string
  floor: number
  type: UnitType
  status: UnitStatus
  createdAt: Date
  updatedAt: Date
}

export type CommonArea = {
  id: number
  buildingId: number
  name: string
  description: string | null
  capacity: number
  isActive: boolean
  createdAt: Date
  updatedAt: Date
}

export type Staff = {
  id: number
  buildingId: number
  userId: number
  user?: User
  role: StaffRole
  startDate: Date
  endDate: Date | null
  createdAt: Date
  updatedAt: Date
}